[orchestration-engine] - Removed some initializers that consumers may find confusing - TT
